% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_draw.R
\name{random_draw}
\alias{random_draw}
\title{Random integer draw Function}
\usage{
random_draw(x, y)
}
\arguments{
\item{x}{A lower_bound number}

\item{y}{A upper_bound number}
}
\description{
This function returns an integer randomly draw between two user-defined bounds
}
\examples{
random_draw()
}

